requirements.txt
aiohttp==3.9.1
aiofiles==23.2.0
beautifulsoup4==4.12.2
redis==5.0.1
cassandra-driver==3.28.0
pyyaml==6.0.1
urllib3==2.1.0
robotparser==0.1.0
prometheus-client==0.19.0
asyncio==3.4.3
lxml==4.9.3
hashlib
logging
argparse
datetime
json
time
re
urllib.parse
uuid
collections
dataclasses
typing
pathlib